PF-2198 Create custom BiOp Annual Report results page * refactoring class/file location to be appropriate
